Use FileInstanceParser interface
